src/main/webapp/app/entities/reservation/condominium/condominium-update.component.ts
import { Component, Vue, Inject } from 'vue-property-decorator';

import { required } from 'vuelidate/lib/validators';

import AlertService from '@/shared/alert/alert.service';

import { ICondominium, Condominium } from '@/shared/model/reservation/condominium.model';
import CondominiumService from './condominium.service';

const validations: any = {
  condominium: {
    name: {
      required,
    },
  },
};

@Component({
  validations,
})
export default class CondominiumUpdate extends Vue {
  @Inject('condominiumService') private condominiumService: () => CondominiumService;
  @Inject('alertService') private alertService: () => AlertService;

  public condominium: ICondominium = new Condominium();
  public isSaving = false;
  public currentLanguage = '';

  beforeRouteEnter(to, from, next) {
    next(vm => {
      if (to.params.condominiumId) {
        vm.retrieveCondominium(to.params.condominiumId);
      }
    });
  }

  created(): void {
    this.currentLanguage = this.$store.getters.currentLanguage;
    this.$store.watch(
      () => this.$store.getters.currentLanguage,
      () => {
        this.currentLanguage = this.$store.getters.currentLanguage;
      }
    );
  }

  public save(): void {
    this.isSaving = true;
    if (this.condominium.id) {
      this.condominiumService()
        .update(this.condominium)
        .then(param => {
          this.isSaving = false;
          this.$router.go(-1);
          const message = this.$t('reservationApp.reservationCondominium.updated', { param: param.id });
          return this.$root.$bvToast.toast(message.toString(), {
            toaster: 'b-toaster-top-center',
            title: 'Info',
            variant: 'info',
            solid: true,
            autoHideDelay: 5000,
          });
        })
        .catch(error => {
          this.isSaving = false;
          this.alertService().showHttpError(this, error.response);
        });
    } else {
      this.condominiumService()
        .create(this.condominium)
        .then(param => {
          this.isSaving = false;
          this.$router.go(-1);
          const message = this.$t('reservationApp.reservationCondominium.created', { param: param.id });
          this.$root.$bvToast.toast(message.toString(), {
            toaster: 'b-toaster-top-center',
            title: 'Success',
            variant: 'success',
            solid: true,
            autoHideDelay: 5000,
          });
        })
        .catch(error => {
          this.isSaving = false;
          this.alertService().showHttpError(this, error.response);
        });
    }
  }

  public retrieveCondominium(condominiumId): void {
    this.condominiumService()
      .find(condominiumId)
      .then(res => {
        this.condominium = res;
      })
      .catch(error => {
        this.alertService().showHttpError(this, error.response);
      });
  }

  public previousState(): void {
    this.$router.go(-1);
  }

  public initRelationships(): void {}
}
